// packages/frontend/src/generated/types/LoginResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { UserResponse } from './UserResponse.ts'

/**
 * @description Functional DTO returned by authentication login endpoints.
 */
export interface LoginResponse {
  /**
   * @type string
   */
  accessToken: string
  /**
   * @type string
   */
  refreshToken: string
  /**
   * @description Response DTO for the `user` entity.
   * @type object
   */
  user: UserResponse
}